src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
anyio==3.6.1
    # via httpcore
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
h11==0.12.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.15.0
    # via httpx
httpx[http2]==0.23.0
    # via -r requirements.in
hyperframe==6.0.1
    # via h2
idna==3.3
    # via
    #   anyio
    #   rfc3986
rfc3986[idna2008]==1.5.0
    # via httpx
ruamel-yaml==0.17.21
    # via -r requirements.in
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typing-extensions==4.2.0
    # via anyio
